+++ src/ReparacoesLN/SSReparacoes/PlanoTrabalho.java
@@ -85,7 +85,7 @@ public class PlanoTrabalho {
 		return this.passos.stream().map(PassoReparacao::getMateriais).collect(Collectors.toList());
 	}
 
-	public CustoTotalReparacao getPrecoEfetivo(Integer precoHora) {
+	public CustoTotalReparacao getPrecoEfetivo(Double precoHora) {
 		CustoTotalReparacao ctr = new CustoTotalReparacao(precoHora);
 		for (PassoReparacao p : passosRealizados){
 			p.getCusto(ctr);
